Address feedback

+++ nemo/collections/asr/models/clustering_diarizer.py
@@ -72,7 +72,7 @@ def get_available_model_names(class_name):
     return list(map(lambda x: x.pretrained_model_name, available_models))
 
 
-class ClusteringDiarizer(Model, DiarizationMixin, torch.nn.Module):
+class ClusteringDiarizer(torch.nn.Module, Model, DiarizationMixin):
     """
     Inference model Class for offline speaker diarization. 
     This class handles required functionality for diarization : Speech Activity Detection, Segmentation, 
@@ -117,14 +117,14 @@ class ClusteringDiarizer(Model, DiarizationMixin, torch.nn.Module):
         model_path = self._cfg.diarizer.vad.model_path
         if model_path.endswith('.nemo'):
             self._vad_model = EncDecClassificationModel.restore_from(model_path, map_location=self._cfg.map_location)
-            self.log_info("VAD model loaded locally from {}".format(model_path))
+            logging.info("VAD model loaded locally from {}".format(model_path))
         else:
             if model_path not in get_available_model_names(EncDecClassificationModel):
                 logging.warning(
                     "requested {} model name not available in pretrained models, instead".format(model_path)
                 )
                 model_path = "vad_telephony_marblenet"
-            self.log_info("Loading pretrained {} model from NGC".format(model_path))
+            logging.info("Loading pretrained {} model from NGC".format(model_path))
             self._vad_model = EncDecClassificationModel.from_pretrained(
                 model_name=model_path, map_location=self._cfg.map_location
             )
@@ -144,19 +144,19 @@ class ClusteringDiarizer(Model, DiarizationMixin, torch.nn.Module):
                 self._speaker_model = EncDecSpeakerLabelModel.restore_from(
                     model_path, map_location=self._cfg.map_location
                 )
-                self.log_info("Speaker Model restored locally from {}".format(model_path))
+                logging.info("Speaker Model restored locally from {}".format(model_path))
             elif model_path.endswith('.ckpt'):
                 self._speaker_model = EncDecSpeakerLabelModel.load_from_checkpoint(
                     model_path, map_location=self._cfg.map_location
                 )
-                self.log_info("Speaker Model restored locally from {}".format(model_path))
+                logging.info("Speaker Model restored locally from {}".format(model_path))
             else:
                 if model_path not in get_available_model_names(EncDecSpeakerLabelModel):
                     logging.warning(
                         "requested {} model name not available in pretrained models, instead".format(model_path)
                     )
                     model_path = "ecapa_tdnn"
-                self.log_info("Loading pretrained {} model from NGC".format(model_path))
+                logging.info("Loading pretrained {} model from NGC".format(model_path))
                 self._speaker_model = EncDecSpeakerLabelModel.from_pretrained(
                     model_name=model_path, map_location=self._cfg.map_location
                 )
@@ -249,7 +249,7 @@ class ClusteringDiarizer(Model, DiarizationMixin, torch.nn.Module):
         else:
             # Generate predictions with overlapping input segments. Then a smoothing filter is applied to decide the label for a frame spanned by multiple segments.
             # smoothing_method would be either in majority vote (median) or average (mean)
-            self.log_info("Generating predictions with overlapping input segments")
+            logging.info("Generating predictions with overlapping input segments")
             smoothing_pred_dir = generate_overlap_vad_seq(
                 frame_pred_dir=self._vad_dir,
                 smoothing_method=self._vad_params.smoothing,
@@ -261,7 +261,7 @@ class ClusteringDiarizer(Model, DiarizationMixin, torch.nn.Module):
             self.vad_pred_dir = smoothing_pred_dir
             frame_length_in_sec = 0.01
 
-        self.log_info("Converting frame level prediction to speech/no-speech segment in start and end times format.")
+        logging.info("Converting frame level prediction to speech/no-speech segment in start and end times format.")
 
         vad_params = self._vad_params if isinstance(self._vad_params, (DictConfig, dict)) else self._vad_params.dict()
         table_out_dir = generate_vad_segment_table(
@@ -286,7 +286,7 @@ class ClusteringDiarizer(Model, DiarizationMixin, torch.nn.Module):
     def _run_segmentation(self, window: float, shift: float, scale_tag: str = ''):
 
         self.subsegments_manifest_path = os.path.join(self._speaker_dir, f'subsegments{scale_tag}.json')
-        self.log_info(
+        logging.info(
             f"Subsegmentation for embedding extraction:{scale_tag.replace('_',' ')}, {self.subsegments_manifest_path}"
         )
         self.subsegments_manifest_path = segments_manifest_to_subsegments_manifest(
@@ -308,7 +308,7 @@ class ClusteringDiarizer(Model, DiarizationMixin, torch.nn.Module):
             manifest_vad_input = self._diarizer_params.manifest_filepath
 
             if self._auto_split:
-                self.log_info("Split long audio file to avoid CUDA memory issue")
+                logging.info("Split long audio file to avoid CUDA memory issue")
                 logging.debug("Try smaller split_duration if you still have CUDA memory issue")
                 config = {
                     'input': manifest_vad_input,
@@ -342,7 +342,7 @@ class ClusteringDiarizer(Model, DiarizationMixin, torch.nn.Module):
         This method extracts speaker embeddings from segments passed through manifest_file
         Optionally you may save the intermediate speaker embeddings for debugging or any use. 
         """
-        self.log_info("Extracting embeddings for Diarization")
+        logging.info("Extracting embeddings for Diarization")
         self._setup_spkr_test_data(manifest_file)
         self.embeddings = {}
         self._speaker_model.eval()
@@ -388,7 +388,7 @@ class ClusteringDiarizer(Model, DiarizationMixin, torch.nn.Module):
             name = os.path.join(embedding_dir, prefix)
             self._embeddings_file = name + f'_embeddings.pkl'
             pkl.dump(self.embeddings, open(self._embeddings_file, 'wb'))
-            self.log_info("Saved embedding files to {}".format(embedding_dir))
+            logging.info("Saved embedding files to {}".format(embedding_dir))
 
     def path2audio_files_to_manifest(self, paths2audio_files, manifest_filepath):
         with open(manifest_filepath, 'w', encoding='utf-8') as fp:
@@ -408,7 +408,10 @@ class ClusteringDiarizer(Model, DiarizationMixin, torch.nn.Module):
         self._out_dir = self._diarizer_params.out_dir
 
         self._speaker_dir = os.path.join(self._diarizer_params.out_dir, 'speaker_outputs')
-        shutil.rmtree(self._speaker_dir, ignore_errors=True)
+
+        if os.path.exists(self._speaker_dir):
+            logging.warning("Deleting previous clustering diarizer outputs.")
+            shutil.rmtree(self._speaker_dir, ignore_errors=True)
         os.makedirs(self._speaker_dir)
 
         if not os.path.exists(self._out_dir):
@@ -460,7 +463,7 @@ class ClusteringDiarizer(Model, DiarizationMixin, torch.nn.Module):
             device=self._speaker_model.device,
             verbose=self.verbose,
         )
-        self.log_info("Outputs are saved in {} directory".format(os.path.abspath(self._diarizer_params.out_dir)))
+        logging.info("Outputs are saved in {} directory".format(os.path.abspath(self._diarizer_params.out_dir)))
 
         # Scoring
         return score_labels(
@@ -556,7 +559,3 @@ class ClusteringDiarizer(Model, DiarizationMixin, torch.nn.Module):
     @property
     def verbose(self) -> bool:
         return self._cfg.verbose
-
-    def log_info(self, *args, **kwargs):
-        if self.verbose:
-            logging.info(*args, **kwargs)
